+++ server/src/server.ts
@@ -10,15 +10,14 @@ import {
 	RequestHandler, NotificationHandler,
 	Diagnostic, DiagnosticSeverity, Range, Files, CancellationToken,
 	TextDocuments, TextDocument, TextDocumentSyncKind, TextEdit, TextDocumentIdentifier, TextDocumentSaveReason,
-	Command, WorkspaceChange,
 	CodeActionRequest, VersionedTextDocumentIdentifier,
 	ExecuteCommandRequest, DidChangeWatchedFilesNotification, DidChangeConfigurationNotification,
 	Proposed, ProposedFeatures
 } from 'vscode-languageserver';
 
 import Uri from 'vscode-uri';
-import path = require('path');
 import fs = require('fs');
+import * as path from 'path';
 
 namespace Is {
 	const toString = Object.prototype.toString;
@@ -99,6 +98,7 @@ namespace DirectoryItem {
 
 interface TextDocumentSettings {
 	validate: boolean;
+	packageManager: 'npm' | 'yarn';
 	autoFix: boolean;
 	autoFixOnSave: boolean;
 	options: any | undefined;
@@ -107,6 +107,7 @@ interface TextDocumentSettings {
 	workspaceFolder: Proposed.WorkspaceFolder | undefined;
 	workingDirectory: DirectoryItem | undefined;
 	library: ESLintModule | undefined;
+	resolvedGlobalPackageManagerPath: string | undefined;
 }
 
 interface ESLintAutoFixEdit {
@@ -291,8 +292,32 @@ let connection = createConnection(ProposedFeatures.all);
 let documents: TextDocuments = new TextDocuments();
 let projectDocuments: TextDocument[] = null;
 
-let globalNodePath: string = undefined;
-
+let _globalNpmPath: string | null | undefined;
+function globalNpmPath(): string {
+	if (_globalNpmPath === void 0) {
+		_globalNpmPath = Files.resolveGlobalNodePath(trace);
+		if (_globalNpmPath === void 0) {
+			_globalNpmPath = null;
+		}
+	}
+	if (_globalNpmPath === null) {
+		return undefined;
+	}
+	return _globalNpmPath;
+}
+let _globalYarnPath: string | undefined;
+function globalYarnPath(): string {
+	if (_globalYarnPath === void 0) {
+		_globalYarnPath = Files.resolveGlobalYarnPath(trace);
+		if (_globalYarnPath === void 0) {
+			_globalYarnPath = null;
+		}
+	}
+	if (_globalYarnPath === null) {
+		return undefined;
+	}
+	return _globalYarnPath;
+}
 let path2Library: Map<string, ESLintModule> = new Map<string, ESLintModule>();
 let document2Settings: Map<string, Thenable<TextDocumentSettings>> = new Map<string, Thenable<TextDocumentSettings>>();
 
@@ -303,20 +328,32 @@ function resolveSettings(document: TextDocument): Thenable<TextDocumentSettings>
 		return resultPromise;
 	}
 	resultPromise = connection.workspace.getConfiguration({ scopeUri: uri, section: '' }).then((settings: TextDocumentSettings) => {
+		if (settings.packageManager === 'npm') {
+			settings.resolvedGlobalPackageManagerPath = globalNpmPath();
+		} else if (settings.packageManager === 'yarn') {
+			settings.resolvedGlobalPackageManagerPath = globalYarnPath();
+		}
 		let uri = Uri.parse(document.uri);
 		let promise: Thenable<string>
 		if (uri.scheme === 'file') {
 			let file = uri.fsPath;
 			let directory = path.dirname(file);
 			if (settings.nodePath) {
-				promise = Files.resolve('eslint', settings.nodePath, settings.nodePath, trace).then<string, string>(undefined, () => {
-					return Files.resolve('eslint', globalNodePath, directory, trace);
+				let nodePath = settings.nodePath;
+				if (!path.isAbsolute(nodePath) && settings.workspaceFolder !== void 0) {
+					let uri = Uri.parse(settings.workspaceFolder.uri);
+					if (uri.scheme === 'file') {
+						nodePath = path.join(uri.fsPath, nodePath);
+					}
+				}
+				promise = Files.resolve('eslint', nodePath, nodePath, trace).then<string, string>(undefined, () => {
+					return Files.resolve('eslint', settings.resolvedGlobalPackageManagerPath, directory, trace);
 				});
 			} else {
-				promise = Files.resolve('eslint', globalNodePath, directory, trace);
+				promise = Files.resolve('eslint', settings.resolvedGlobalPackageManagerPath, directory, trace);
 			}
 		} else {
-			promise = Files.resolve('eslint', globalNodePath, settings.workspaceFolder ? settings.workspaceFolder.uri : undefined, trace);
+			promise = Files.resolve('eslint', settings.resolvedGlobalPackageManagerPath, settings.workspaceFolder ? settings.workspaceFolder.uri : undefined, trace);
 		}
 		return promise.then((path) => {
 			let library = path2Library.get(path);
@@ -663,7 +700,7 @@ const singleErrorHandlers: ((error: any, document: TextDocument, library: ESLint
 function validateSingle(document: TextDocument, publishDiagnostics: boolean = true): Thenable<void> {
 	// We validate document in a queue but open / close documents directly. So we need to deal with the
 	// fact that a document might be gone from the server.
-	if (!documents && !documents.get(document.uri)) {
+	if (!projectDocuments && !documents.get(document.uri)) {
 		return Promise.resolve(undefined);
 	}
 	return resolveSettings(document).then((settings) => {